+++ ports/esp32s2/common-hal/audiobusio/I2SOut.h
@@ -29,7 +29,7 @@
 #include "supervisor/background_callback.h"
 #include "common-hal/microcontroller/Pin.h"
 
-#include "i2s_common.h"
+#include "common-hal/audiobusio/__init__.h"
 
 // Some boards don't implement I2SOut, so suppress any routines from here.
 #if CIRCUITPY_AUDIOBUSIO_I2SOUT
